+++ api/src/data_migration/transformation/__init__.py
@@ -1,0 +1,6 @@
+from typing import TypeAlias
+
+from src.db.models.staging.forecast import Tforecast, TforecastHist
+from src.db.models.staging.synopsis import Tsynopsis, TsynopsisHist
+
+SourceSummary: TypeAlias = Tforecast | Tsynopsis | TforecastHist | TsynopsisHist
